64ビットから16ビットへの代入が行われていた部分を修正  Fixed assignment from 64-bit to 16-bit

+++ src/Stage/Stage.h
@@ -9,8 +9,8 @@ namespace stage{
     class Stage{
         private:
         game::Game* m_game;
-        uint64_t m_height; // 高さ
-        uint64_t m_width; // 幅
+        uint16_t m_height; // 高さ
+        uint16_t m_width; // 幅
         collision::Rectangle m_right; // 右の壁の当たり判定  Right wall collision detection
         collision::Rectangle m_bottom; // 下の壁の当たり判定  Bottom wall collision detection
         collision::Rectangle m_left; // 左の壁の当たり判定  Left wall collision detection
